requirements.txt
# ecofil API — Python Dependencies
# Only production dependencies are listed here.
# Install with:  pip install -r requirements.txt
#
# To regenerate after adding a new package:
#   pip freeze | grep -i "package_name"

# --- Web Framework ---
fastapi==0.135.3
uvicorn[standard]==0.44.0
gunicorn==24.0.0
starlette==1.0.0
python-multipart==0.0.26

# --- Database (PostgreSQL + Async) ---
SQLAlchemy==2.0.49
alembic==1.18.4
greenlet==3.4.0
asyncpg>=0.30.0
psycopg2-binary>=2.9.9

# --- Authentication & Security ---
PyJWT==2.12.1
bcrypt==5.0.0

# --- Data Validation ---
pydantic==2.12.5
pydantic_core==2.41.5
email-validator==2.3.0
annotated-types==0.7.0

# --- HTTP Client (TMDB API — async) ---
httpx==0.28.1
idna==3.11
certifi==2026.2.25

# --- System Metadata Collection ---
psutil==7.2.2

# --- Configuration ---
python-dotenv==1.2.2

# --- Template Engine (Alembic dependency) ---
Mako==1.3.10
MarkupSafe==3.0.3

# --- Testing ---
pytest==9.0.3
pytest-asyncio==1.3.0
httpcore==1.0.9
anyio==4.13.0

# --- Misc Runtime ---
click==8.3.2
h11==0.16.0
sniffio>=1.3.0
typing_extensions==4.15.0
pluggy==1.6.0
iniconfig==2.3.0
packaging==26.0
dnspython==2.8.0
# --- AI & General Intelligence ---
google-genai
groq==0.18.0
slowapi==0.1.9
google-auth==2.38.0
cloudinary==1.42.1
